Rewrite of symbol creation code. Partial fix for 1201325.
Symbols cannot have 'tranform' attribute. The old code copied the group
'transform' attribute to the symbol. The new code creates an intermediate
group with the required 'transform' if it is not a simple translation. A
simple translation is removed and applied instead to the new <use> element.

Multiple objects can now be turned into a symbol without first grouping the
objects. (At the cost that multiple groups can no longer be turned into
multiple symbols at one go.) This better matches marker and pattern creation.